Fix analytics plan policy test import
