+++ templates/analytics/server/plugins/agent-chat.ts
@@ -4,7 +4,6 @@ import {
   getRequestRunContext,
   loadActionsFromStaticRegistry,
   type AgentLoopFinalResponseGuardContext,
-  type ActionEntry,
 } from "@agent-native/core/server";
 
 import actionsRegistry from "../../.generated/actions-registry.js";
@@ -27,6 +26,7 @@ import {
   listScopedSettingRecords,
   resolveSettingsScope,
 } from "../lib/scoped-settings";
+import { applyAnalyticsPlanModePolicy } from "./agent-chat-plan-mode";
 
 const DATA_DICT_PREFIX = "data-dict-";
 
@@ -65,33 +65,10 @@ const INITIAL_TOOL_NAMES = [
   "navigate",
 ];
 
-export const PLAN_MODE_ACT_ONLY_TOOLS = new Set([
-  "query-agent-native-analytics",
-  "bigquery",
-  "provider-api-request",
-  "provider-corpus-job",
-  "query-staged-dataset",
-  "account-deep-dive",
-  "hubspot-deals",
-  "hubspot-records",
-  "gong-calls",
-  "jira-search",
-  "slack-messages",
-  "sentry",
-]);
-
-export function applyAnalyticsPlanModePolicy(
-  actions: Record<string, ActionEntry>,
-): Record<string, ActionEntry> {
-  return Object.fromEntries(
-    Object.entries(actions).map(([name, entry]) => [
-      name,
-      PLAN_MODE_ACT_ONLY_TOOLS.has(name)
-        ? { ...entry, allowInPlanMode: false }
-        : entry,
-    ]),
-  );
-}
+export {
+  applyAnalyticsPlanModePolicy,
+  PLAN_MODE_ACT_ONLY_TOOLS,
+} from "./agent-chat-plan-mode";
 
 function latestUserText(
   messages: AgentLoopFinalResponseGuardContext["messages"],
